Added contacts-schemas and validateBody
