+++ decorators/validateBody.js
@@ -1,0 +1,14 @@
+import { HttpError } from '../helpers/index.js';
+
+const validateBody = (schema) => {
+  const func = (req, res, next) => {
+    const { error } = schema.validate(req.body);
+    if (error) {
+      next(HttpError(400, error.message));
+    }
+    next();
+  };
+  return func;
+};
+
+export default validateBody;
