+++ routes/api/contacts-router.js
@@ -1,6 +1,8 @@
 import express from 'express';
 
 import contactsController from '../../controllers/contacts-controllers.js';
+import contactsSchema from '../../schemas/contacts-schemas.js';
+import { validateBody } from '../../decorators/index.js';
 
 const router = express.Router();
 
@@ -8,10 +10,18 @@ router.get('/', contactsController.getContactsList);
 
 router.get('/:contactId', contactsController.getContactById);
 
-router.post('/', contactsController.addNewContact);
+router.post(
+  '/',
+  validateBody(contactsSchema.contactsAddSchema),
+  contactsController.addNewContact
+);
 
 router.delete('/:contactId', contactsController.deleteContactById);
 
-router.put('/:contactId', contactsController.updateContactById);
+router.put(
+  '/:contactId',
+  validateBody(contactsSchema.contactsAddSchema),
+  contactsController.updateContactById
+);
 
 export default router;
